requirements.txt
aiohttp==3.11.11
google-genai
tqdm==4.67.1
pandas>=2.0.0
seaborn>=0.13.0
matplotlib>=3.8.0
numpy>=1.24.0
